+++ rio_tiler/io/stac.py
@@ -1,11 +1,12 @@
 """rio_tiler.io.stac: STAC reader."""
 
 import json
+import math
 import os
 import warnings
 from collections.abc import Iterator, Sequence
 from threading import Lock
-from typing import Any
+from typing import Any, TypedDict
 from urllib.parse import urlparse
 
 import attr
@@ -16,13 +17,15 @@ from affine import Affine
 from cachetools import LRUCache, cached
 from cachetools.keys import hashkey
 from morecantile import TileMatrixSet
-from rasterio.transform import array_bounds
+from rasterio.crs import CRS
+from rasterio.features import bounds as featureBounds
+from rasterio.transform import array_bounds, from_bounds
 
 from rio_tiler.constants import WEB_MERCATOR_TMS, WGS84_CRS
 from rio_tiler.errors import InvalidAssetName, MissingAssets
 from rio_tiler.io.base import BaseReader, MultiBaseReader
 from rio_tiler.io.rasterio import Reader
-from rio_tiler.types import AssetInfo
+from rio_tiler.types import AssetInfo, AssetType
 
 try:
     from boto3.session import Session as boto3_session
@@ -194,6 +197,128 @@ def _to_pystac_item(item: dict | pystac.Item | None) -> pystac.Item | None:
     return item
 
 
+class Projection(TypedDict):
+    """Projection info extracted from STAC Item."""
+
+    width: int
+    height: int
+    bounds: tuple[float, float, float, float]
+    transform: Affine
+    crs: CRS
+
+
+def _extract_proj_info(
+    item: pystac.Item, assets: list[str] | None = None
+) -> Projection | None:
+    """Extract projection info from STAC Item.
+
+    Args:
+        item (pystac.Item): STAC Item.
+        assets (list[str]): An optional list of assets to use instead of the item's proj value.
+
+    Returns:
+        Projection: projection info extracted from the STAC Item.
+
+    """
+    if not item.ext.has("proj"):
+        return None
+
+    crs = item.ext.proj.crs_string
+
+    # Asset level projection extension
+    assets = assets or list(item.assets)
+    asset_proj: dict[str, Any] = {}
+
+    for asset in assets:
+        asset_info = item.assets[asset]
+        if asset_info.ext.has("proj"):
+            if all(
+                [
+                    asset_info.ext.proj.transform,
+                    asset_info.ext.proj.shape,
+                ]
+            ):
+                tr = Affine(*asset_info.ext.proj.transform)
+                asset_proj[asset] = {
+                    "shape": asset_info.ext.proj.shape,
+                    "transform": tr,
+                    "bounds": array_bounds(
+                        asset_info.ext.proj.shape[0],
+                        asset_info.ext.proj.shape[1],
+                        tr,
+                    ),
+                    "crs_string": asset_info.ext.proj.crs_string or crs,
+                }
+
+    if asset_proj:
+        # 1. check that all assets have the same projection info
+        # if multiple CRS if will be too expensive to handle
+        crss = [p["crs_string"] for p in asset_proj.values()]
+        if crs:
+            crss.append(crs)
+
+        if len(set(crss)) == 1:
+            crs = next(iter(crss))
+            if not crs:
+                return None
+
+            # 2. create unified bounds, transform and shape for the item based on the assets info
+            bounds = (
+                min(p["bounds"][0] for p in asset_proj.values()),
+                min(p["bounds"][1] for p in asset_proj.values()),
+                max(p["bounds"][2] for p in asset_proj.values()),
+                max(p["bounds"][3] for p in asset_proj.values()),
+            )
+
+            # 3. Get the highest resolution asset and use its transform to calculate
+            # the theoritical shape of the item
+            highest_res_asset = min(
+                asset_proj.items(),
+                key=lambda p: (
+                    abs(p[1]["transform"][0]),  # pixel width
+                    abs(p[1]["transform"][4]),  # pixel height
+                ),
+            )
+            xres = highest_res_asset[1]["transform"][0]
+            yres = highest_res_asset[1]["transform"][4]
+
+            # 4. Get dataset shape from bounds and resolution
+            height = abs(math.floor((bounds[3] - bounds[1]) / yres))
+            width = abs(math.floor((bounds[2] - bounds[0]) / xres))
+
+            transform = from_bounds(*bounds, width, height)
+
+            return Projection(
+                width=width,
+                height=height,
+                bounds=bounds,
+                transform=transform,
+                crs=CRS.from_string(crs),
+            )
+
+    # Item Level projection extension
+    if all(
+        [
+            item.ext.proj.transform,
+            item.ext.proj.shape,
+            crs,
+        ]
+    ):
+        height, width = item.ext.proj.shape
+        transform = Affine(*item.ext.proj.transform)
+        bounds = array_bounds(height, width, transform)
+
+        return Projection(
+            width=width,
+            height=height,
+            bounds=bounds,
+            transform=transform,
+            crs=CRS.from_string(crs),
+        )
+
+    return None
+
+
 @attr.s
 class STACReader(MultiBaseReader):
     """STAC Reader.
@@ -208,7 +333,7 @@ class STACReader(MultiBaseReader):
         exclude_assets (set of string, optional): Exclude specific assets.
         include_asset_types (set of string, optional): Only include some assets base on their type.
         exclude_asset_types (set of string, optional): Exclude some assets base on their type.
-        default_assets (list of string, optional): Default assets to use if none are defined.
+        default_assets (list of string or dict, optional): Default assets to use if none are defined.
         reader (rio_tiler.io.BaseReader, optional): rio-tiler Reader. Defaults to `rio_tiler.io.Reader`.
         reader_options (dict, optional): Additional option to forward to the Reader. Defaults to `{}`.
         fetch_options (dict, optional): Options to pass to `rio_tiler.io.stac.fetch` function fetching the STAC Items. Defaults to `{}`.
@@ -246,43 +371,40 @@ class STACReader(MultiBaseReader):
     exclude_asset_types: set[str] | None = attr.ib(default=None)
 
     assets: Sequence[str] = attr.ib(init=False)
-    default_assets: Sequence[str] | None = attr.ib(default=None)
+    default_assets: Sequence[AssetType] | None = attr.ib(default=None)
 
     reader: type[BaseReader] = attr.ib(default=Reader)
     reader_options: dict = attr.ib(factory=dict)
 
     fetch_options: dict = attr.ib(factory=dict)
 
-    ctx: rasterio.Env = attr.ib(default=rasterio.Env)
+    ctx: type[rasterio.Env] = attr.ib(default=rasterio.Env)
 
     def __attrs_post_init__(self):
         """Fetch STAC Item and get list of valid assets."""
         self.item = self.item or pystac.Item.from_dict(
             fetch(self.input, **self.fetch_options), self.input
         )
-
-        self.bounds = tuple(self.item.bbox)
-        self.crs = WGS84_CRS
-
-        if hasattr(self.item, "ext") and self.item.ext.has("proj"):
-            if all(
-                [
-                    self.item.ext.proj.transform,
-                    self.item.ext.proj.shape,
-                    self.item.ext.proj.crs_string,
-                ]
-            ):
-                self.height, self.width = self.item.ext.proj.shape
-                self.transform = Affine(*self.item.ext.proj.transform)
-                self.bounds = array_bounds(self.height, self.width, self.transform)
-                self.crs = rasterio.crs.CRS.from_string(self.item.ext.proj.crs_string)
-
-        self.minzoom = self.minzoom if self.minzoom is not None else self._minzoom
-        self.maxzoom = self.maxzoom if self.maxzoom is not None else self._maxzoom
-
         self.assets = self.get_asset_list()
         if not self.assets:
             raise MissingAssets("No valid asset found. Asset's media types not supported")
+
+        if proj := _extract_proj_info(self.item, assets=self.assets):
+            self.height = proj["height"]
+            self.width = proj["width"]
+            self.bounds = proj["bounds"]
+            self.transform = proj["transform"]
+            self.crs = proj["crs"]
+        else:
+            self.bounds = (
+                tuple(self.item.bbox)
+                if self.item.bbox
+                else featureBounds(self.item.geometry)
+            )
+            self.crs = WGS84_CRS
+
+        self.minzoom = self.minzoom if self.minzoom is not None else self._minzoom
+        self.maxzoom = self.maxzoom if self.maxzoom is not None else self._maxzoom
 
     def get_asset_list(self) -> list[str]:
         """Get valid asset list"""
@@ -317,48 +439,52 @@ class STACReader(MultiBaseReader):
 
         return asset, None
 
-    def _get_asset_info(self, asset: str) -> AssetInfo:
+    def _get_asset_info(self, asset: AssetType) -> AssetInfo:
         """Validate asset names and return asset's info.
 
         Args:
-            asset (str): STAC asset name.
+            asset (AssetType): STAC asset name.
 
         Returns:
             AssetInfo: STAC asset info.
 
         """
-        asset, vrt_options = self._parse_vrt_asset(asset)
+        asset_name: str
+        if isinstance(asset, dict):
+            if not asset.get("name"):
+                raise ValueError("asset dictionary does not have `name` key")
+            asset_name = asset["name"]
+        else:
+            asset_name = asset
+
+        asset_name, vrt_options = self._parse_vrt_asset(asset_name)
+
+        if asset_name not in self.assets:
+            raise InvalidAssetName(
+                f"'{asset_name}' is not valid, should be one of {self.assets}"
+            )
+
+        asset_info = self.item.assets[asset_name]
+        extras = asset_info.extra_fields
 
         method_options: dict[str, Any] = {}
+        reader_options: dict[str, Any] = {}
+        if isinstance(asset, dict):
+            if indexes := asset.get("indexes"):
+                method_options["indexes"] = indexes
+            if expr := asset.get("expression"):
+                method_options["expression"] = expr
 
-        # NOTE: asset can be in form of
-        # "{asset_name}|some_option=some_value&another_option=another_value"
-        if "|" in asset:
-            asset, params = asset.split("|", 1)
-            # NOTE: Construct method options from params
-            if params:
-                for param in params.split("&"):
-                    key, value = param.split("=", 1)
-                    if key == "indexes":
-                        method_options["indexes"] = list(map(int, value.split(",")))
-                    elif key == "expression":
-                        method_options["expression"] = value
-
-        if asset not in self.assets:
-            raise InvalidAssetName(
-                f"'{asset}' is not valid, should be one of {self.assets}"
-            )
+            # TODO: handle `bands` options
+            # convert bands to indexes based on the band metadata
 
         asset_modified = "expression" in method_options or vrt_options
 
-        asset_info = self.item.assets[asset]
-        extras = asset_info.extra_fields
-
         info = AssetInfo(
             url=asset_info.get_absolute_href() or asset_info.href,
-            name=asset,
+            name=asset_name,
             media_type=asset_info.media_type,
-            reader_options={},
+            reader_options=reader_options,
             method_options=method_options,
         )